style(user): change user id type 'Integer' into 'Long'

+++ src/main/java/org/standard/dreamcalendar/domain/user/UserRepository.java
@@ -5,12 +5,11 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.lang.NonNull;
 import org.springframework.transaction.annotation.Transactional;
-import org.standard.dreamcalendar.domain.user.model.User;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     @Transactional
     @Modifying(clearAutomatically = true)
     @Query("update User u set u.accessToken = ?2 where u.id = ?1")
